requirements.txt
clearml==1.15.1
hydra-core==1.3.2
numpy==1.24.4
omegaconf==2.3.0
pandas==1.5.3
pytorch_lightning==2.2.3
torch==2.3.0
replay-rec==0.18.1
transformers==4.49.0
tqdm==4.67.1

# visualization
matplotlib==3.7.2
streamlit==1.43.2
